add blog reading page, make author card into link, fix issues from schema change

+++ app/api/blogs/[blogId]/route.ts
@@ -19,6 +19,10 @@ export async function DELETE(
 
   const { blogId } = params;
 
+  if (!blogId || typeof blogId !== "string") {
+    throw new Error("Invalid blod id");
+  }
+
   const blog = await prisma.blog.delete({
     where: {
       id: blogId,
